backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel    
python-jose[cryptography]
pytest
httpx
pytest-asyncio
python-dotenv
bcrypt==3.2.2
passlib[bcrypt]==1.7.4

